refactor(dashboard): extract data, tasks, and persistence logic to hooks
host status removed from display in host page
